XFree86 4.3.99.902 (RC 2)

+++ di/options.c
@@ -46,7 +46,7 @@ in this Software without prior written authorization from The Open Group.
  * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
  */
-/* $XFree86: xc/programs/lbxproxy/di/options.c,v 1.8 2003/09/13 21:33:10 dawes Exp $ */
+/* $XFree86: xc/programs/lbxproxy/di/options.c,v 1.9 2003/12/19 02:05:39 dawes Exp $ */
 
 #include <stdio.h>
 #include "X.h"
@@ -682,7 +682,7 @@ LbxLookupBitmapCompMethod (server, methodOpCode)
     {
 	LbxBitmapCompMethod *method;
 
-	method = &server->LbxBitmapCompMethods[server->lbxNegOpt.bitmapCompMethods[i]];
+	method = &server->LbxBitmapCompMethods[(int)(server->lbxNegOpt.bitmapCompMethods[i])];
 
 	if (method->methodOpCode == methodOpCode)
 	    return (method);
@@ -704,7 +704,7 @@ LbxLookupPixmapCompMethod (server, methodOpCode)
     {
 	LbxPixmapCompMethod *method;
 
-	method = &server->LbxPixmapCompMethods[server->lbxNegOpt.pixmapCompMethods[i]];
+	method = &server->LbxPixmapCompMethods[(int)(server->lbxNegOpt.pixmapCompMethods[i])];
 
 	if (method->methodOpCode == methodOpCode)
 	    return (method);
@@ -722,7 +722,7 @@ LbxFindPreferredBitmapCompMethod (server)
     if (server->lbxNegOpt.numBitmapCompMethods == 0)
 	return NULL;
     else
-	return (&server->LbxBitmapCompMethods[server->lbxNegOpt.bitmapCompMethods[0]]);
+	return (&server->LbxBitmapCompMethods[(int)(server->lbxNegOpt.bitmapCompMethods[0])]);
 }
 
 
@@ -742,7 +742,7 @@ LbxFindPreferredPixmapCompMethod (server, format, depth)
 
 	for (i = 0; i < server->lbxNegOpt.numPixmapCompMethods; i++)
 	{
-	    method = &server->LbxPixmapCompMethods[server->lbxNegOpt.pixmapCompMethods[i]];
+	    method = &server->LbxPixmapCompMethods[(int)(server->lbxNegOpt.pixmapCompMethods[i])];
 
 	    if ((method->formatMask & (1 << format)))
 	    {
